# contracts/http/identity.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D3 Identity API",
    "version": "1.0.0",
    "description": "Local email/password identity: registration, login, rotating refresh sessions, logout, and authenticated profile (D3-ID-001)."
  },
  "x-requirements": ["D3-ID-001", "D3-SEC-001"],
  "paths": {
    "/v1/auth/register": {
      "post": {
        "summary": "Register a local account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {"schema": {"$ref": "#/components/schemas/RegisterRequest"}}
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegisterResponse"}}}
          },
          "400": {"$ref": "#/components/responses/Error"},
          "409": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "summary": "Sign in and open a refresh session",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Access token and rotating refresh token",
            "headers": {
              "Set-Cookie": {"schema": {"type": "string"}, "description": "HttpOnly SameSite=Lax D3_REFRESH_TOKEN cookie"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}
          },
          "400": {"$ref": "#/components/responses/Error"},
          "401": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "summary": "Rotate the refresh session and mint a new access token",
        "parameters": [{"$ref": "#/components/parameters/RefreshCookie"}],
        "responses": {
          "200": {
            "description": "New access token and rotated refresh token",
            "headers": {
              "Set-Cookie": {"schema": {"type": "string"}, "description": "Rotated HttpOnly SameSite=Lax D3_REFRESH_TOKEN cookie"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}
          },
          "401": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "summary": "Revoke a refresh session",
        "parameters": [{"name": "D3_REFRESH_TOKEN", "in": "cookie", "required": false, "schema": {"type": "string"}}],
        "responses": {
          "204": {
            "description": "Session revoked when present",
            "headers": {
              "Set-Cookie": {"schema": {"type": "string"}, "description": "Expired D3_REFRESH_TOKEN cookie"}
            }
          }
        }
      }
    },
    "/v1/users/me": {
      "get": {
        "summary": "The authenticated caller's public profile",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Profile",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProfileResponse"}}}
          },
          "401": {"$ref": "#/components/responses/Error"},
          "404": {"$ref": "#/components/responses/Error"}
        }
      },
      "patch": {
        "summary": "Update the authenticated caller's public profile",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateProfileRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProfileResponse"}}}
          },
          "400": {"$ref": "#/components/responses/Error"},
          "401": {"$ref": "#/components/responses/Error"},
          "404": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/.well-known/jwks.json": {
      "get": {
        "summary": "Public signing keys for access-token verification",
        "responses": {
          "200": {
            "description": "JSON Web Key Set",
            "content": {"application/json": {"schema": {"type": "object"}}}
          }
        }
      }
    },
    "/internal/v1/service-tokens": {
      "post": {
        "summary": "Issue a short-lived Judge service token to Battle",
        "security": [{"serviceClient": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {"schema": {"$ref": "#/components/schemas/ServiceTokenRequest"}}
          }
        },
        "responses": {
          "200": {
            "description": "Short-lived Battle service token",
            "content": {
              "application/json": {"schema": {"$ref": "#/components/schemas/ServiceTokenResponse"}}
            }
          },
          "400": {"$ref": "#/components/responses/Error"},
          "401": {"$ref": "#/components/responses/Error"}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"},
      "serviceClient": {"type": "http", "scheme": "basic", "description": "Battle service client credentials"}
    },
    "parameters": {
      "RefreshCookie": {
        "name": "D3_REFRESH_TOKEN",
        "in": "cookie",
        "required": true,
        "schema": {"type": "string"}
      }
    },
    "responses": {
      "Error": {
        "description": "Versioned error",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "handle", "displayName", "password"],
        "properties": {
          "email": {"type": "string", "format": "email", "description": "Stored lowercased; matching is case-insensitive."},
          "handle": {"type": "string", "pattern": "^[a-z0-9-]{1,39}$"},
          "displayName": {"type": "string", "maxLength": 80},
          "password": {"type": "string", "minLength": 8, "maxLength": 200}
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {"type": "string", "format": "email"},
          "password": {"type": "string"}
        }
      },
      "RegisterResponse": {
        "type": "object",
        "required": ["userId"],
        "properties": {"userId": {"type": "string", "format": "uuid"}}
      },
      "TokenResponse": {
        "type": "object",
        "required": ["userId", "accessToken"],
        "properties": {
          "userId": {"type": "string", "format": "uuid"},
          "accessToken": {"type": "string"}
        }
      },
      "ProfileResponse": {
        "type": "object",
        "required": ["userId", "handle", "email", "displayName"],
        "properties": {
          "userId": {"type": "string", "format": "uuid"},
          "handle": {"type": "string"},
          "email": {"type": "string", "format": "email"},
          "displayName": {"type": "string"}
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "required": ["displayName"],
        "properties": {
          "displayName": {"type": "string", "maxLength": 80}
        }
      },
      "ServiceTokenRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["scope"],
        "properties": {
          "scope": {"type": "string", "enum": ["judge.submit", "judge.read"]}
        }
      },
      "ServiceTokenResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["accessToken", "tokenType", "expiresIn"],
        "properties": {
          "accessToken": {"type": "string"},
          "tokenType": {"type": "string", "const": "Bearer"},
          "expiresIn": {"type": "integer", "minimum": 1, "maximum": 900}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message", "correlationId"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "correlationId": {"type": "string"}
        }
      }
    }
  }
}